+++ CHANGELOG.md
@@ -4,14 +4,17 @@
 * Added the config interface `experimental` to group experimental features.
 * Added a flag `enablePreviousNameRecording` to add previous event/view and current view names as segmentation (Experimental!)
 * Added a flag `enableVisibilityTracking` to add app visibility info to views (Experimental!)
+* Added `Content` feature methods:
+  * `enterContentZone`, to start Content checks(Experimental!)
+  * `exitContentZone`, to stop content checks (Experimental!)
 * Added support for `List` values in user given segmentations of timed events.
-
-* Updated the underlying Firebase Messaging SDK to version 24.0.3
 
 * Mitigated an issue where an event was not recorded if a `count` was not provided.
 * Fixed an issue where automatic crash reporting failed to capture Flutter framework errors when using the newly introduced config option.
 * Addressed an issue where asynchronous Dart errors were not being captured.
 * Addressed an issue that prevented the stacktrace from being properly recognized on the server
+
+* Updated the underlying Firebase Messaging SDK to version 24.0.3
 
 * Updated underlying Android SDK version to 24.7.4
 * Updated underlying iOS SDK version to 24.7.7